+++ library/src/main/java/com/franmontiel/localechanger/AppLocaleChanger.java
@@ -19,7 +19,7 @@ package com.franmontiel.localechanger;
 import android.content.Context;
 import android.content.res.Configuration;
 import android.os.Build;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.RequiresApi;
 
 import java.util.Locale;