+++ library/src/main/java/com/franmontiel/localechanger/LocaleChanger.java
@@ -25,7 +25,6 @@ import com.franmontiel.localechanger.utils.SystemLocaleRetriever;
 import java.util.List;
 import java.util.Locale;
 
-
 public class LocaleChanger {
 
     private static LocaleChangerDelegate delegate;
@@ -34,11 +33,13 @@ public class LocaleChanger {
     }
 
     /**
-     * Initialize the LocaleChanger, this method needs to be called before calling any other method and it must be called only once.
+     * Initialize the LocaleChanger, this method needs to be called before calling any other method.
      * <p>
      * If this method was never invoked before it sets a Locale from the supported list if a language match is found with the system Locales,
      * if no match is found the first Locale in the list will be set.<p>
      * If this method was invoked before it will load a Locale previously set.
+     * <p>
+     * If this method is invoked when the library is already initialized the new settings will be applied from now on.
      *
      * @param context
      * @param supportedLocales a list of your app supported Locales
@@ -52,23 +53,22 @@ public class LocaleChanger {
     }
 
     /**
-     * Initialize the LocaleChanger, this method needs to be called before calling any other method and it must be called only once.<p>
+     * Initialize the LocaleChanger, this method needs to be called before calling any other method.
+     * <p>
      * If this method was never invoked before it sets a Locale resolved using the provided {@link MatchingAlgorithm} and {@link LocalePreference}.<p>
      * If this method was invoked before it will load a Locale previously set.
+     * <p>
+     * If this method is invoked when the library is already initialized the new settings will be applied from now on.
      *
      * @param context
      * @param supportedLocales  a list of your app supported Locales
      * @param matchingAlgorithm used to find a match between supported and system Locales
      * @param preference        used to indicate what Locale is preferred to load in case of a match
-     * @throws IllegalStateException if the LocaleChanger is already initialized
      */
     public static void initialize(Context context,
                                   List<Locale> supportedLocales,
                                   MatchingAlgorithm matchingAlgorithm,
                                   LocalePreference preference) {
-        if (delegate != null) {
-            throw new IllegalStateException("LocaleChanger is already initialized");
-        }
 
         delegate = new LocaleChangerDelegate(
                 new LocalePersistor(context),
@@ -82,11 +82,17 @@ public class LocaleChanger {
         delegate.initialize();
     }
 
+    private static void checkInitialization() {
+        if (delegate == null)
+            throw new IllegalStateException("LocaleChanger is not initialized. Please first call LocaleChanger.initialize");
+    }
+
     /**
      * Clears any Locale set and resolve and load a new default one.
      * This method can be useful if the app implements new supported Locales and it is needed to reload the default one in case there is a best match.
      */
     public static void resetLocale() {
+        checkInitialization();
         delegate.resetLocale();
     }
 
@@ -96,6 +102,7 @@ public class LocaleChanger {
      * @param supportedLocale a supported Locale that will be used to resolve the Locale to set.
      */
     public static void setLocale(Locale supportedLocale) {
+        checkInitialization();
         delegate.setLocale(supportedLocale);
     }
 
@@ -105,6 +112,7 @@ public class LocaleChanger {
      * @return
      */
     public static Locale getLocale() {
+        checkInitialization();
         return delegate.getLocale();
     }
 
@@ -116,6 +124,7 @@ public class LocaleChanger {
      * @return the resulting context that should be provided to the super method call.
      */
     public static Context configureBaseContext(Context context) {
+        checkInitialization();
         return delegate.configureBaseContext(context);
     }
 
@@ -123,6 +132,7 @@ public class LocaleChanger {
      * This method should be called from Application#onConfigurationChanged()
      */
     public static void onConfigurationChanged() {
+        checkInitialization();
         delegate.onConfigurationChanged();
     }
 }