+++ library/src/main/java/com/franmontiel/localechanger/LocalePersistor.java
@@ -18,8 +18,8 @@ package com.franmontiel.localechanger;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.util.Locale;