+++ library/src/main/java/com/franmontiel/localechanger/utils/ActivityRecreationHelper.java
@@ -19,8 +19,8 @@ package com.franmontiel.localechanger.utils;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.app.ActivityOptionsCompat;
+import androidx.core.app.ActivityCompat;
+import androidx.core.app.ActivityOptionsCompat;
 
 import java.util.HashMap;
 import java.util.Locale;