+++ sample/src/androidTest/java/com/franmontiel/localechanger/sample/SampleActivityTest.java
@@ -16,16 +16,14 @@
 
 package com.franmontiel.localechanger.sample;
 
-import android.support.test.espresso.Espresso;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.espresso.Espresso;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import com.franmontiel.localechanger.sample.pageobjects.SampleScreen;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-
-import java.util.Locale;
 
 /**
  * Advice: It is recommended to run the tests with a system locale configured not listed in the supported ones (to avoid false positives).