+++ sample/src/main/java/com/franmontiel/localechanger/sample/SampleFragment.java
@@ -18,8 +18,8 @@ package com.franmontiel.localechanger.sample;
 
 import android.app.DatePickerDialog;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;