+++ sample/src/main/java/com/franmontiel/localechanger/sample/SampleFragmentContainerActivity.java
@@ -19,8 +19,8 @@ package com.franmontiel.localechanger.sample;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.franmontiel.localechanger.LocaleChanger;