+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+Version 1.0 *(2020-02-13)*
+----------------------------
+ * Library migrated to AndroidX
+ * Changed behavior on initialize method. Now it can be called multiple times, in that case the settings will be updated.
+ * IllegalStateException is now thrown when the library is not initialized, prior to that it was NullPointerException.
 
 Version 0.9.2 *(2018-01-26)*
 ----------------------------